+++ Tests/DependencyInjection/GosWebSocketExtensionTest.php
@@ -52,33 +52,6 @@ class GosWebSocketExtensionTest extends AbstractExtensionTestCase
         $this->load();
     }
 
-    public function testContainerIsLoadedWithEnvVars(): void
-    {
-        $this->container->setParameter(
-            'kernel.bundles',
-            [
-                'GosPubSubRouterBundle' => GosPubSubRouterBundle::class,
-                'GosWebSocketBundle' => GosWebSocketBundle::class,
-            ]
-        );
-
-        $this->container->setParameter('env(GOS_WEB_SOCKET_SERVER_HOST)', '127.0.0.1');
-        $this->container->setParameter('env(GOS_WEB_SOCKET_SERVER_PORT)', 8080);
-
-        $bundleConfig = [
-            'server' => [
-                'host' => '%env(GOS_WEB_SOCKET_SERVER_HOST)%',
-                'port' => '%env(int:GOS_WEB_SOCKET_SERVER_PORT)%',
-                'origin_check' => false,
-            ],
-        ];
-
-        $this->load($bundleConfig);
-
-        $this->assertContainerBuilderHasParameter('gos_web_socket.server.port');
-        $this->assertContainerBuilderHasParameter('gos_web_socket.server.host');
-    }
-
     public function testContainerIsLoadedWithPubSubBundleIntegration(): void
     {
         $this->container->setParameter(
@@ -100,25 +73,9 @@ class GosWebSocketExtensionTest extends AbstractExtensionTestCase
             ],
         ];
 
-        // Prepend config now to allow the prepend pass to work
-        $this->container->prependExtensionConfig('gos_web_socket', $bundleConfig);
-
-        // Also load the bundle config so it is passed to the extension load method
         $this->load($bundleConfig);
 
-        $this->assertSame(
-            [
-                [
-                    'routers' => [
-                        'websocket' => [
-                            'resources' => [],
-                        ],
-                    ],
-                ],
-            ],
-            $this->container->getExtensionConfig('gos_pubsub_router'),
-            'The GosPubSubRouterBundle should be configured when able.'
-        );
+        $this->assertContainerBuilderHasParameter('gos_web_socket.router_resources', []);
     }
 
     public function testContainerIsLoadedWithPubSubBundleIntegrationAndLegacyConfiguration(): void
@@ -150,26 +107,13 @@ class GosWebSocketExtensionTest extends AbstractExtensionTestCase
             ],
         ];
 
-        // Prepend config now to allow the prepend pass to work
-        $this->container->prependExtensionConfig('gos_web_socket', $bundleConfig);
-
-        // Also load the bundle config so it is passed to the extension load method
         $this->load($bundleConfig);
 
-        $this->assertSame(
+        $this->assertContainerBuilderHasParameter(
+            'gos_web_socket.router_resources',
             [
-                [
-                    'routers' => [
-                        'websocket' => [
-                            'resources' => [
-                                'example.yaml',
-                            ],
-                        ],
-                    ],
-                ],
-            ],
-            $this->container->getExtensionConfig('gos_pubsub_router'),
-            'The GosPubSubRouterBundle should be configured when able.'
+                'example.yaml',
+            ]
         );
     }
 
@@ -205,26 +149,13 @@ class GosWebSocketExtensionTest extends AbstractExtensionTestCase
             ],
         ];
 
-        // Prepend config now to allow the prepend pass to work
-        $this->container->prependExtensionConfig('gos_web_socket', $bundleConfig);
-
-        // Also load the bundle config so it is passed to the extension load method
         $this->load($bundleConfig);
 
-        $this->assertSame(
+        $this->assertContainerBuilderHasParameter(
+            'gos_web_socket.router_resources',
             [
-                [
-                    'routers' => [
-                        'websocket' => [
-                            'resources' => [
-                                'example.yaml',
-                            ],
-                        ],
-                    ],
-                ],
-            ],
-            $this->container->getExtensionConfig('gos_pubsub_router'),
-            'The GosPubSubRouterBundle should be configured when able.'
+                'example.yaml',
+            ]
         );
     }
 
@@ -257,29 +188,16 @@ class GosWebSocketExtensionTest extends AbstractExtensionTestCase
             ],
         ];
 
-        // Prepend config now to allow the prepend pass to work
-        $this->container->prependExtensionConfig('gos_web_socket', $bundleConfig);
-
-        // Also load the bundle config so it is passed to the extension load method
         $this->load($bundleConfig);
 
-        $this->assertSame(
+        $this->assertContainerBuilderHasParameter(
+            'gos_web_socket.router_resources',
             [
                 [
-                    'routers' => [
-                        'websocket' => [
-                            'resources' => [
-                                [
-                                    'resource' => 'example.yaml',
-                                    'type' => null,
-                                ],
-                            ],
-                        ],
-                    ],
+                    'resource' => 'example.yaml',
+                    'type' => null,
                 ],
-            ],
-            $this->container->getExtensionConfig('gos_pubsub_router'),
-            'The GosPubSubRouterBundle should be configured when able.'
+            ]
         );
     }
 
@@ -315,29 +233,16 @@ class GosWebSocketExtensionTest extends AbstractExtensionTestCase
             ],
         ];
 
-        // Prepend config now to allow the prepend pass to work
-        $this->container->prependExtensionConfig('gos_web_socket', $bundleConfig);
-
-        // Also load the bundle config so it is passed to the extension load method
         $this->load($bundleConfig);
 
-        $this->assertSame(
+        $this->assertContainerBuilderHasParameter(
+            'gos_web_socket.router_resources',
             [
                 [
-                    'routers' => [
-                        'websocket' => [
-                            'resources' => [
-                                [
-                                    'resource' => 'example.yaml',
-                                    'type' => null,
-                                ],
-                            ],
-                        ],
-                    ],
+                    'resource' => 'example.yaml',
+                    'type' => null,
                 ],
-            ],
-            $this->container->getExtensionConfig('gos_pubsub_router'),
-            'The GosPubSubRouterBundle should be configured when able.'
+            ]
         );
     }
 
@@ -500,15 +405,12 @@ class GosWebSocketExtensionTest extends AbstractExtensionTestCase
             ]
         );
 
-        $this->container->setParameter('env(GOS_WEB_SOCKET_WAMP_PUSHER_HOST)', '127.0.0.1');
-        $this->container->setParameter('env(GOS_WEB_SOCKET_WAMP_PUSHER_PORT)', 1337);
-
         $bundleConfig = [
             'pushers' => [
                 'wamp' => [
                     'enabled' => true,
-                    'host' => '%env(GOS_WEB_SOCKET_WAMP_PUSHER_HOST)%',
-                    'port' => '%env(int:GOS_WEB_SOCKET_WAMP_PUSHER_PORT)%',
+                    'host' => '127.0.0.1',
+                    'port' => 1337,
                     'ssl' => false,
                     'origin' => null,
                 ],